Remove unused files from Repo Update naming of certain files
